requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
newspaper4k
lxml_html_clean
arxiv
wikipedia
googlesearch-python
pypdf
pycountry
pygame
matplotlib
scipy